sessiontest: split test for improvement ci
