+++ session/sessiontest/session_test.go
@@ -18,9 +18,7 @@ import (
 	"context"
 	"fmt"
 	"net"
-	"strconv"
 	"strings"
-	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -28,7 +26,6 @@ import (
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/domain"
-	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/ast"
@@ -37,7 +34,6 @@ import (
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/parser/terror"
-	plannercore "github.com/pingcap/tidb/planner/core"
 	"github.com/pingcap/tidb/privilege/privileges"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx"
@@ -52,8 +48,6 @@ import (
 	"github.com/pingcap/tidb/util/memory"
 	"github.com/pingcap/tidb/util/sqlexec"
 	"github.com/stretchr/testify/require"
-	"github.com/tikv/client-go/v2/tikv"
-	"github.com/tikv/client-go/v2/txnkv/transaction"
 )
 
 func TestSchemaCheckerSQL(t *testing.T) {
@@ -132,142 +126,6 @@ func TestSchemaCheckerSQL(t *testing.T) {
 	require.Error(t, tk.ExecToErr(`commit;`))
 }
 
-func TestDisableTxnAutoRetry(t *testing.T) {
-	store := testkit.CreateMockStoreWithSchemaLease(t, 1*time.Second)
-
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk1 := testkit.NewTestKit(t, store)
-	tk2 := testkit.NewTestKit(t, store)
-
-	tk1.MustExec("use test")
-	tk2.MustExec("use test")
-
-	tk1.MustExec("create table no_retry (id int)")
-	tk1.MustExec("insert into no_retry values (1)")
-	tk1.MustExec("set @@tidb_disable_txn_auto_retry = 1")
-
-	tk1.MustExec("begin")
-	tk1.MustExec("update no_retry set id = 2")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update no_retry set id = 3")
-	tk2.MustExec("commit")
-
-	// No auto retry because tidb_disable_txn_auto_retry is set to 1.
-	_, err := tk1.Session().Execute(context.Background(), "commit")
-	require.Error(t, err)
-
-	// session 1 starts a transaction early.
-	// execute a select statement to clear retry history.
-	tk1.MustExec("select 1")
-	err = tk1.Session().PrepareTxnCtx(context.Background())
-	require.NoError(t, err)
-	// session 2 update the value.
-	tk2.MustExec("update no_retry set id = 4")
-	// AutoCommit update will retry, so it would not fail.
-	tk1.MustExec("update no_retry set id = 5")
-
-	// RestrictedSQL should retry.
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnOthers)
-	tk1.Session().ExecuteInternal(ctx, "begin")
-
-	tk2.MustExec("update no_retry set id = 6")
-
-	tk1.Session().ExecuteInternal(ctx, "update no_retry set id = 7")
-	tk1.Session().ExecuteInternal(ctx, "commit")
-
-	// test for disable transaction local latch
-	defer config.RestoreFunc()()
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.TxnLocalLatches.Enabled = false
-	})
-	tk1.MustExec("begin")
-	tk1.MustExec("update no_retry set id = 9")
-
-	tk2.MustExec("update no_retry set id = 8")
-
-	_, err = tk1.Session().Execute(context.Background(), "commit")
-	require.Error(t, err)
-	require.True(t, kv.ErrWriteConflict.Equal(err), fmt.Sprintf("err %v", err))
-	require.Contains(t, err.Error(), kv.TxnRetryableMark)
-	tk1.MustExec("rollback")
-
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.TxnLocalLatches.Enabled = true
-	})
-	tk1.MustExec("begin")
-	tk2.MustExec("alter table no_retry add index idx(id)")
-	tk2.MustQuery("select * from no_retry").Check(testkit.Rows("8"))
-	tk1.MustExec("update no_retry set id = 10")
-	_, err = tk1.Session().Execute(context.Background(), "commit")
-	require.Error(t, err)
-
-	// set autocommit to begin and commit
-	tk1.MustExec("set autocommit = 0")
-	tk1.MustQuery("select * from no_retry").Check(testkit.Rows("8"))
-	tk2.MustExec("update no_retry set id = 11")
-	tk1.MustExec("update no_retry set id = 12")
-	_, err = tk1.Session().Execute(context.Background(), "set autocommit = 1")
-	require.Error(t, err)
-	require.True(t, kv.ErrWriteConflict.Equal(err), fmt.Sprintf("err %v", err))
-	require.Contains(t, err.Error(), kv.TxnRetryableMark)
-	tk1.MustExec("rollback")
-	tk2.MustQuery("select * from no_retry").Check(testkit.Rows("11"))
-
-	tk1.MustExec("set autocommit = 0")
-	tk1.MustQuery("select * from no_retry").Check(testkit.Rows("11"))
-	tk2.MustExec("update no_retry set id = 13")
-	tk1.MustExec("update no_retry set id = 14")
-	_, err = tk1.Session().Execute(context.Background(), "commit")
-	require.Error(t, err)
-	require.True(t, kv.ErrWriteConflict.Equal(err), fmt.Sprintf("err %v", err))
-	require.Contains(t, err.Error(), kv.TxnRetryableMark)
-	tk1.MustExec("rollback")
-	tk2.MustQuery("select * from no_retry").Check(testkit.Rows("13"))
-}
-
-// The Read-only flags are checked in the planning stage of queries,
-// but this test checks we check them again at commit time.
-// The main use case for this is a long-running auto-commit statement.
-func TestAutoCommitRespectsReadOnly(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	var wg sync.WaitGroup
-	tk1 := testkit.NewTestKit(t, store)
-	tk2 := testkit.NewTestKit(t, store)
-	require.NoError(t, tk1.Session().Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil, nil))
-	require.NoError(t, tk2.Session().Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil, nil))
-
-	tk1.MustExec("create table test.auto_commit_test (a int)")
-	wg.Add(1)
-	go func() {
-		err := tk1.ExecToErr("INSERT INTO test.auto_commit_test VALUES (SLEEP(1))")
-		require.True(t, terror.ErrorEqual(err, plannercore.ErrSQLInReadOnlyMode), fmt.Sprintf("err %v", err))
-		wg.Done()
-	}()
-	tk2.MustExec("SET GLOBAL tidb_restricted_read_only = 1")
-	err := tk2.ExecToErr("INSERT INTO test.auto_commit_test VALUES (0)") // should also be an error
-	require.True(t, terror.ErrorEqual(err, plannercore.ErrSQLInReadOnlyMode), fmt.Sprintf("err %v", err))
-	// Reset and check with the privilege to ignore the readonly flag and continue to insert.
-	wg.Wait()
-	tk1.MustExec("SET GLOBAL tidb_restricted_read_only = 0")
-	tk1.MustExec("SET GLOBAL tidb_super_read_only = 0")
-	tk1.MustExec("GRANT RESTRICTED_REPLICA_WRITER_ADMIN on *.* to 'root'")
-
-	wg.Add(1)
-	go func() {
-		tk1.MustExec("INSERT INTO test.auto_commit_test VALUES (SLEEP(1))")
-		wg.Done()
-	}()
-	tk2.MustExec("SET GLOBAL tidb_restricted_read_only = 1")
-	tk2.MustExec("INSERT INTO test.auto_commit_test VALUES (0)")
-
-	// wait for go routines
-	wg.Wait()
-	tk1.MustExec("SET GLOBAL tidb_restricted_read_only = 0")
-	tk1.MustExec("SET GLOBAL tidb_super_read_only = 0")
-}
-
 func TestLoadSchemaFailed(t *testing.T) {
 	originalRetryTime := domain.SchemaOutOfDateRetryTimes.Load()
 	originalRetryInterval := domain.SchemaOutOfDateRetryInterval.Load()
@@ -334,58 +192,6 @@ func TestSysdateIsNow(t *testing.T) {
 	tk.MustExec("set @@tidb_sysdate_is_now=true")
 	tk.MustQuery("show variables like '%tidb_sysdate_is_now%'").Check(testkit.Rows("tidb_sysdate_is_now ON"))
 	require.True(t, tk.Session().GetSessionVars().SysdateIsNow)
-}
-
-func TestEnableLegacyInstanceScope(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-
-	// enable 'switching' to SESSION variables
-	tk.MustExec("set tidb_enable_legacy_instance_scope = 1")
-	tk.MustExec("set tidb_general_log = 1")
-	tk.MustQuery(`show warnings`).Check(testkit.Rows(fmt.Sprintf("Warning %d modifying tidb_general_log will require SET GLOBAL in a future version of TiDB", errno.ErrInstanceScope)))
-	require.True(t, tk.Session().GetSessionVars().EnableLegacyInstanceScope)
-
-	// disable 'switching' to SESSION variables
-	tk.MustExec("set tidb_enable_legacy_instance_scope = 0")
-	tk.MustGetErrCode("set tidb_general_log = 1", errno.ErrGlobalVariable)
-	require.False(t, tk.Session().GetSessionVars().EnableLegacyInstanceScope)
-}
-
-func TestSetPDClientDynamicOption(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("0"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = 0.5;")
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("0.5"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = 1;")
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("1"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = 1.5;")
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("1.5"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = 10;")
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("10"))
-	require.Error(t, tk.ExecToErr("set tidb_tso_client_batch_max_wait_time = 0;"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = -1;")
-	tk.MustQuery("show warnings").Check(testkit.RowsWithSep("|", "Warning|1292|Truncated incorrect tidb_tso_client_batch_max_wait_time value: '-1'"))
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("0"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = -0.1;")
-	tk.MustQuery("show warnings").Check(testkit.RowsWithSep("|", "Warning|1292|Truncated incorrect tidb_tso_client_batch_max_wait_time value: '-0.1'"))
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("0"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = 10.1;")
-	tk.MustQuery("show warnings").Check(testkit.RowsWithSep("|", "Warning|1292|Truncated incorrect tidb_tso_client_batch_max_wait_time value: '10.1'"))
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("10"))
-	tk.MustExec("set global tidb_tso_client_batch_max_wait_time = 11;")
-	tk.MustQuery("show warnings").Check(testkit.RowsWithSep("|", "Warning|1292|Truncated incorrect tidb_tso_client_batch_max_wait_time value: '11'"))
-	tk.MustQuery("select @@tidb_tso_client_batch_max_wait_time;").Check(testkit.Rows("10"))
-
-	tk.MustQuery("select @@tidb_enable_tso_follower_proxy;").Check(testkit.Rows("0"))
-	tk.MustExec("set global tidb_enable_tso_follower_proxy = on;")
-	tk.MustQuery("select @@tidb_enable_tso_follower_proxy;").Check(testkit.Rows("1"))
-	tk.MustExec("set global tidb_enable_tso_follower_proxy = off;")
-	tk.MustQuery("select @@tidb_enable_tso_follower_proxy;").Check(testkit.Rows("0"))
-	require.Error(t, tk.ExecToErr("set tidb_tso_client_batch_max_wait_time = 0;"))
 }
 
 func TestWriteOnMultipleCachedTable(t *testing.T) {
@@ -493,190 +299,6 @@ func TestFixSetTiDBSnapshotTS(t *testing.T) {
 	tk.MustExec("use t123")
 }
 
-func TestSetVarHint(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("sql_mode", mysql.DefaultSQLMode))
-	tk.MustQuery("SELECT /*+ SET_VAR(sql_mode=ALLOW_INVALID_DATES) */ @@sql_mode;").Check(testkit.Rows("ALLOW_INVALID_DATES"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@sql_mode;").Check(testkit.Rows(mysql.DefaultSQLMode))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("tmp_table_size", "16777216"))
-	tk.MustQuery("SELECT /*+ SET_VAR(tmp_table_size=1024) */ @@tmp_table_size;").Check(testkit.Rows("1024"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@tmp_table_size;").Check(testkit.Rows("16777216"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("range_alloc_block_size", "4096"))
-	tk.MustQuery("SELECT /*+ SET_VAR(range_alloc_block_size=4294967295) */ @@range_alloc_block_size;").Check(testkit.Rows("4294967295"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@range_alloc_block_size;").Check(testkit.Rows("4096"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_execution_time", "0"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_execution_time=1) */ @@max_execution_time;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_execution_time;").Check(testkit.Rows("0"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("time_zone", "SYSTEM"))
-	tk.MustQuery("SELECT /*+ SET_VAR(time_zone='+12:00') */ @@time_zone;").Check(testkit.Rows("+12:00"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@time_zone;").Check(testkit.Rows("SYSTEM"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("join_buffer_size", "262144"))
-	tk.MustQuery("SELECT /*+ SET_VAR(join_buffer_size=128) */ @@join_buffer_size;").Check(testkit.Rows("128"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@join_buffer_size;").Check(testkit.Rows("262144"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_length_for_sort_data", "1024"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_length_for_sort_data=4) */ @@max_length_for_sort_data;").Check(testkit.Rows("4"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_length_for_sort_data;").Check(testkit.Rows("1024"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_error_count", "64"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_error_count=0) */ @@max_error_count;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_error_count;").Check(testkit.Rows("64"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("sql_buffer_result", "OFF"))
-	tk.MustQuery("SELECT /*+ SET_VAR(sql_buffer_result=ON) */ @@sql_buffer_result;").Check(testkit.Rows("ON"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@sql_buffer_result;").Check(testkit.Rows("OFF"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_heap_table_size", "16777216"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_heap_table_size=16384) */ @@max_heap_table_size;").Check(testkit.Rows("16384"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_heap_table_size;").Check(testkit.Rows("16777216"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("tmp_table_size", "16777216"))
-	tk.MustQuery("SELECT /*+ SET_VAR(tmp_table_size=16384) */ @@tmp_table_size;").Check(testkit.Rows("16384"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@tmp_table_size;").Check(testkit.Rows("16777216"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("div_precision_increment", "4"))
-	tk.MustQuery("SELECT /*+ SET_VAR(div_precision_increment=0) */ @@div_precision_increment;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@div_precision_increment;").Check(testkit.Rows("4"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("sql_auto_is_null", "OFF"))
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("tidb_enable_noop_functions", "ON"))
-	tk.MustQuery("SELECT /*+ SET_VAR(sql_auto_is_null=1) */ @@sql_auto_is_null;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVarWithoutValidation("tidb_enable_noop_functions", "OFF"))
-	tk.MustQuery("SELECT @@sql_auto_is_null;").Check(testkit.Rows("0"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("sort_buffer_size", "262144"))
-	tk.MustQuery("SELECT /*+ SET_VAR(sort_buffer_size=32768) */ @@sort_buffer_size;").Check(testkit.Rows("32768"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@sort_buffer_size;").Check(testkit.Rows("262144"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_join_size", "18446744073709551615"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_join_size=1) */ @@max_join_size;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_join_size;").Check(testkit.Rows("18446744073709551615"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_seeks_for_key", "18446744073709551615"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_seeks_for_key=1) */ @@max_seeks_for_key;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_seeks_for_key;").Check(testkit.Rows("18446744073709551615"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_sort_length", "1024"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_sort_length=4) */ @@max_sort_length;").Check(testkit.Rows("4"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_sort_length;").Check(testkit.Rows("1024"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("bulk_insert_buffer_size", "8388608"))
-	tk.MustQuery("SELECT /*+ SET_VAR(bulk_insert_buffer_size=0) */ @@bulk_insert_buffer_size;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@bulk_insert_buffer_size;").Check(testkit.Rows("8388608"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("sql_big_selects", "1"))
-	tk.MustQuery("SELECT /*+ SET_VAR(sql_big_selects=0) */ @@sql_big_selects;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@sql_big_selects;").Check(testkit.Rows("1"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("read_rnd_buffer_size", "262144"))
-	tk.MustQuery("SELECT /*+ SET_VAR(read_rnd_buffer_size=1) */ @@read_rnd_buffer_size;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@read_rnd_buffer_size;").Check(testkit.Rows("262144"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("unique_checks", "1"))
-	tk.MustQuery("SELECT /*+ SET_VAR(unique_checks=0) */ @@unique_checks;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@unique_checks;").Check(testkit.Rows("1"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("read_buffer_size", "131072"))
-	tk.MustQuery("SELECT /*+ SET_VAR(read_buffer_size=8192) */ @@read_buffer_size;").Check(testkit.Rows("8192"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@read_buffer_size;").Check(testkit.Rows("131072"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("default_tmp_storage_engine", "InnoDB"))
-	tk.MustQuery("SELECT /*+ SET_VAR(default_tmp_storage_engine='CSV') */ @@default_tmp_storage_engine;").Check(testkit.Rows("CSV"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@default_tmp_storage_engine;").Check(testkit.Rows("InnoDB"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("optimizer_search_depth", "62"))
-	tk.MustQuery("SELECT /*+ SET_VAR(optimizer_search_depth=1) */ @@optimizer_search_depth;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@optimizer_search_depth;").Check(testkit.Rows("62"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("max_points_in_geometry", "65536"))
-	tk.MustQuery("SELECT /*+ SET_VAR(max_points_in_geometry=3) */ @@max_points_in_geometry;").Check(testkit.Rows("3"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@max_points_in_geometry;").Check(testkit.Rows("65536"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("updatable_views_with_limit", "YES"))
-	tk.MustQuery("SELECT /*+ SET_VAR(updatable_views_with_limit=0) */ @@updatable_views_with_limit;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@updatable_views_with_limit;").Check(testkit.Rows("YES"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("optimizer_prune_level", "1"))
-	tk.MustQuery("SELECT /*+ SET_VAR(optimizer_prune_level=0) */ @@optimizer_prune_level;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@optimizer_prune_level;").Check(testkit.Rows("1"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("group_concat_max_len", "1024"))
-	tk.MustQuery("SELECT /*+ SET_VAR(group_concat_max_len=4) */ @@group_concat_max_len;").Check(testkit.Rows("4"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@group_concat_max_len;").Check(testkit.Rows("1024"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("eq_range_index_dive_limit", "200"))
-	tk.MustQuery("SELECT /*+ SET_VAR(eq_range_index_dive_limit=0) */ @@eq_range_index_dive_limit;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@eq_range_index_dive_limit;").Check(testkit.Rows("200"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("sql_safe_updates", "0"))
-	tk.MustQuery("SELECT /*+ SET_VAR(sql_safe_updates=1) */ @@sql_safe_updates;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@sql_safe_updates;").Check(testkit.Rows("0"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("end_markers_in_json", "0"))
-	tk.MustQuery("SELECT /*+ SET_VAR(end_markers_in_json=1) */ @@end_markers_in_json;").Check(testkit.Rows("1"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@end_markers_in_json;").Check(testkit.Rows("0"))
-
-	require.NoError(t, tk.Session().GetSessionVars().SetSystemVar("windowing_use_high_precision", "ON"))
-	tk.MustQuery("SELECT /*+ SET_VAR(windowing_use_high_precision=OFF) */ @@windowing_use_high_precision;").Check(testkit.Rows("0"))
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-	tk.MustQuery("SELECT @@windowing_use_high_precision;").Check(testkit.Rows("1"))
-
-	tk.MustExec("SELECT /*+ SET_VAR(sql_safe_updates = 1) SET_VAR(max_heap_table_size = 1G) */ 1;")
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 0)
-
-	tk.MustExec("SELECT /*+ SET_VAR(collation_server = 'utf8') */ 1;")
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 1)
-	require.EqualError(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings()[0].Err, "[planner:3637]Variable 'collation_server' cannot be set using SET_VAR hint.")
-
-	tk.MustExec("SELECT /*+ SET_VAR(max_size = 1G) */ 1;")
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 1)
-	require.EqualError(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings()[0].Err, "[planner:3128]Unresolved name 'max_size' for SET_VAR hint")
-
-	tk.MustExec("SELECT /*+ SET_VAR(group_concat_max_len = 1024) SET_VAR(group_concat_max_len = 2048) */ 1;")
-	require.Len(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings(), 1)
-	require.EqualError(t, tk.Session().GetSessionVars().StmtCtx.GetWarnings()[0].Err, "[planner:3126]Hint SET_VAR(group_concat_max_len=2048) is ignored as conflicting/duplicated.")
-}
-
 func TestPrepareZero(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
@@ -767,32 +389,6 @@ func TestSetGroupConcatMaxLen(t *testing.T) {
 	// Test illegal type
 	_, err = tk.Exec("set @@group_concat_max_len='hello'")
 	require.True(t, terror.ErrorEqual(err, variable.ErrWrongTypeForVar), fmt.Sprintf("err %v", err))
-}
-
-func TestRetryForCurrentTxn(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	tk.MustExec("create table history (a int)")
-	tk.MustExec("insert history values (1)")
-
-	// Firstly, enable retry.
-	tk.MustExec("set tidb_disable_txn_auto_retry = 0")
-	tk.MustExec("begin")
-	tk.MustExec("update history set a = 2")
-	// Disable retry now.
-	tk.MustExec("set tidb_disable_txn_auto_retry = 1")
-
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("use test")
-	tk1.MustExec("update history set a = 3")
-
-	tk.MustExec("commit")
-	tk.MustQuery("select * from history").Check(testkit.Rows("2"))
 }
 
 // TestTruncateAlloc tests that the auto_increment ID does not reuse the old table's allocator.
@@ -928,213 +524,6 @@ func TestStatementCountLimit(t *testing.T) {
 	tk.MustExec("insert into stmt_count_limit values (3)")
 	_, err = tk.Exec("insert into stmt_count_limit values (4)")
 	require.Error(t, err)
-}
-
-func TestBatchCommit(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_batch_commit = 1")
-	tk.MustExec("set tidb_disable_txn_auto_retry = 0")
-	tk.MustExec("create table t (id int)")
-	defer config.RestoreFunc()()
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.Performance.StmtCountLimit = 3
-	})
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("use test")
-	tk.MustExec("SET SESSION autocommit = 1")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (1)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows())
-	tk.MustExec("insert into t values (2)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows())
-	tk.MustExec("rollback")
-	tk1.MustQuery("select * from t").Check(testkit.Rows())
-
-	// The above rollback will not make the session in transaction.
-	tk.MustExec("insert into t values (1)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("1"))
-	tk.MustExec("delete from t")
-
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (5)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows())
-	tk.MustExec("insert into t values (6)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows())
-	tk.MustExec("insert into t values (7)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("5", "6", "7"))
-
-	// The session is still in transaction.
-	tk.MustExec("insert into t values (8)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("5", "6", "7"))
-	tk.MustExec("insert into t values (9)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("5", "6", "7"))
-	tk.MustExec("insert into t values (10)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("5", "6", "7"))
-	tk.MustExec("commit")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("5", "6", "7", "8", "9", "10"))
-
-	// The above commit will not make the session in transaction.
-	tk.MustExec("insert into t values (11)")
-	tk1.MustQuery("select * from t").Check(testkit.Rows("5", "6", "7", "8", "9", "10", "11"))
-
-	tk.MustExec("delete from t")
-	tk.MustExec("SET SESSION autocommit = 0")
-	tk.MustExec("insert into t values (1)")
-	tk.MustExec("insert into t values (2)")
-	tk.MustExec("insert into t values (3)")
-	tk.MustExec("rollback")
-	tk1.MustExec("insert into t values (4)")
-	tk1.MustExec("insert into t values (5)")
-	tk.MustQuery("select * from t").Check(testkit.Rows("4", "5"))
-}
-
-func TestKVVars(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set @@tidb_backoff_lock_fast = 1")
-	tk.MustExec("set @@tidb_backoff_weight = 100")
-	tk.MustExec("create table if not exists kvvars (a int key)")
-	tk.MustExec("insert into kvvars values (1)")
-	tk.MustExec("begin")
-	txn, err := tk.Session().Txn(false)
-	require.NoError(t, err)
-	vars := txn.GetVars().(*tikv.Variables)
-	require.Equal(t, 1, vars.BackoffLockFast)
-	require.Equal(t, 100, vars.BackOffWeight)
-	tk.MustExec("rollback")
-	tk.MustExec("set @@tidb_backoff_weight = 50")
-	tk.MustExec("set @@autocommit = 0")
-	tk.MustExec("select * from kvvars")
-	require.True(t, tk.Session().GetSessionVars().InTxn())
-	txn, err = tk.Session().Txn(false)
-	require.NoError(t, err)
-	vars = txn.GetVars().(*tikv.Variables)
-	require.Equal(t, 50, vars.BackOffWeight)
-
-	tk.MustExec("set @@autocommit = 1")
-	require.Nil(t, failpoint.Enable("tikvclient/probeSetVars", `return(true)`))
-	tk.MustExec("select * from kvvars where a = 1")
-	require.Nil(t, failpoint.Disable("tikvclient/probeSetVars"))
-	require.True(t, transaction.SetSuccess.Load())
-	transaction.SetSuccess.Store(false)
-}
-
-func TestTxnRetryErrMsg(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk1 := testkit.NewTestKit(t, store)
-	tk2 := testkit.NewTestKit(t, store)
-	tk1.MustExec("use test")
-	tk1.MustExec("create table no_retry (id int)")
-	tk1.MustExec("insert into no_retry values (1)")
-	tk1.MustExec("begin")
-	tk2.MustExec("use test")
-	tk2.MustExec("update no_retry set id = id + 1")
-	tk1.MustExec("update no_retry set id = id + 1")
-	require.NoError(t, failpoint.Enable("tikvclient/mockRetryableErrorResp", `return(true)`))
-	_, err := tk1.Session().Execute(context.Background(), "commit")
-	require.NoError(t, failpoint.Disable("tikvclient/mockRetryableErrorResp"))
-	require.Error(t, err)
-	require.True(t, kv.ErrTxnRetryable.Equal(err), "error: %s", err)
-	require.True(t, strings.Contains(err.Error(), "mock retryable error"), "error: %s", err)
-	require.True(t, strings.Contains(err.Error(), kv.TxnRetryableMark), "error: %s", err)
-}
-
-func TestSetTxnScope(t *testing.T) {
-	// Check the default value of @@tidb_enable_local_txn and @@txn_scope whitout configuring the zone label.
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustQuery("select @@global.tidb_enable_local_txn;").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Check the default value of @@tidb_enable_local_txn and @@txn_scope with configuring the zone label.
-	require.NoError(t, failpoint.Enable("tikvclient/injectTxnScope", `return("bj")`))
-	tk = testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustQuery("select @@global.tidb_enable_local_txn;").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	require.NoError(t, failpoint.Disable("tikvclient/injectTxnScope"))
-
-	// @@tidb_enable_local_txn is off without configuring the zone label.
-	tk = testkit.NewTestKit(t, store)
-	tk.MustQuery("select @@global.tidb_enable_local_txn;").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to local.
-	err := tk.ExecToErr("set @@txn_scope = 'local';")
-	require.Error(t, err)
-	require.Regexp(t, `.*txn_scope can not be set to local when tidb_enable_local_txn is off.*`, err)
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to global.
-	tk.MustExec("set @@txn_scope = 'global';")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-
-	// @@tidb_enable_local_txn is off with configuring the zone label.
-	require.NoError(t, failpoint.Enable("tikvclient/injectTxnScope", `return("bj")`))
-	tk = testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustQuery("select @@global.tidb_enable_local_txn;").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to local.
-	err = tk.ExecToErr("set @@txn_scope = 'local';")
-	require.Error(t, err)
-	require.Regexp(t, `.*txn_scope can not be set to local when tidb_enable_local_txn is off.*`, err)
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to global.
-	tk.MustExec("set @@txn_scope = 'global';")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	require.NoError(t, failpoint.Disable("tikvclient/injectTxnScope"))
-
-	// @@tidb_enable_local_txn is on without configuring the zone label.
-	tk = testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set global tidb_enable_local_txn = on;")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to local.
-	err = tk.ExecToErr("set @@txn_scope = 'local';")
-	require.Error(t, err)
-	require.Regexp(t, `.*txn_scope can not be set to local when zone label is empty or "global".*`, err)
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to global.
-	tk.MustExec("set @@txn_scope = 'global';")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-
-	// @@tidb_enable_local_txn is on with configuring the zone label.
-	require.NoError(t, failpoint.Enable("tikvclient/injectTxnScope", `return("bj")`))
-	tk = testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set global tidb_enable_local_txn = on;")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.LocalTxnScope))
-	require.Equal(t, "bj", tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to global.
-	tk.MustExec("set @@txn_scope = 'global';")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.GlobalTxnScope))
-	require.Equal(t, kv.GlobalTxnScope, tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Set @@txn_scope to local.
-	tk.MustExec("set @@txn_scope = 'local';")
-	tk.MustQuery("select @@txn_scope;").Check(testkit.Rows(kv.LocalTxnScope))
-	require.Equal(t, "bj", tk.Session().GetSessionVars().CheckAndGetTxnScope())
-	// Try to set @@txn_scope to an invalid value.
-	err = tk.ExecToErr("set @@txn_scope='foo'")
-	require.Error(t, err)
-	require.Regexp(t, `.*txn_scope value should be global or local.*`, err)
-	require.NoError(t, failpoint.Disable("tikvclient/injectTxnScope"))
 }
 
 func TestDoDDLJobQuit(t *testing.T) {
@@ -1316,29 +705,6 @@ func TestIssue21943(t *testing.T) {
 	require.Equal(t, "[variable:1238]Variable 'last_plan_from_cache' is a read only variable", err.Error())
 }
 
-func TestRemovedSysVars(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	variable.RegisterSysVar(&variable.SysVar{Scope: variable.ScopeGlobal | variable.ScopeSession, Name: "bogus_var", Value: "acdc"})
-	result := tk.MustQuery("SHOW GLOBAL VARIABLES LIKE 'bogus_var'")
-	result.Check(testkit.Rows("bogus_var acdc"))
-	result = tk.MustQuery("SELECT @@GLOBAL.bogus_var")
-	result.Check(testkit.Rows("acdc"))
-	tk.MustExec("SET GLOBAL bogus_var = 'newvalue'")
-
-	// unregister
-	variable.UnregisterSysVar("bogus_var")
-
-	result = tk.MustQuery("SHOW GLOBAL VARIABLES LIKE 'bogus_var'")
-	result.Check(testkit.Rows()) // empty
-	_, err := tk.Exec("SET GLOBAL bogus_var = 'newvalue'")
-	require.Equal(t, "[variable:1193]Unknown system variable 'bogus_var'", err.Error())
-	_, err = tk.Exec("SELECT @@GLOBAL.bogus_var")
-	require.Equal(t, "[variable:1193]Unknown system variable 'bogus_var'", err.Error())
-}
-
 func TestCorrectScopeError(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 	tk := testkit.NewTestKit(t, store)
@@ -1376,58 +742,6 @@ func TestCorrectScopeError(t *testing.T) {
 	variable.UnregisterSysVar("sv_global")
 	variable.UnregisterSysVar("sv_session")
 	variable.UnregisterSysVar("sv_both")
-}
-
-func TestTiKVSystemVars(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	result := tk.MustQuery("SHOW GLOBAL VARIABLES LIKE 'tidb_gc_enable'") // default is on from the sysvar
-	result.Check(testkit.Rows("tidb_gc_enable ON"))
-	result = tk.MustQuery("SELECT variable_value FROM mysql.tidb WHERE variable_name = 'tikv_gc_enable'")
-	result.Check(testkit.Rows()) // but no value in the table (yet) because the value has not been set and the GC has never been run
-
-	// update will set a value in the table
-	tk.MustExec("SET GLOBAL tidb_gc_enable = 1")
-	result = tk.MustQuery("SELECT variable_value FROM mysql.tidb WHERE variable_name = 'tikv_gc_enable'")
-	result.Check(testkit.Rows("true"))
-
-	tk.MustExec("UPDATE mysql.tidb SET variable_value = 'false' WHERE variable_name='tikv_gc_enable'")
-	result = tk.MustQuery("SELECT @@tidb_gc_enable;")
-	result.Check(testkit.Rows("0")) // reads from mysql.tidb value and changes to false
-
-	tk.MustExec("SET GLOBAL tidb_gc_concurrency = -1") // sets auto concurrency and concurrency
-	result = tk.MustQuery("SELECT variable_value FROM mysql.tidb WHERE variable_name = 'tikv_gc_auto_concurrency'")
-	result.Check(testkit.Rows("true"))
-	result = tk.MustQuery("SELECT variable_value FROM mysql.tidb WHERE variable_name = 'tikv_gc_concurrency'")
-	result.Check(testkit.Rows("-1"))
-
-	tk.MustExec("SET GLOBAL tidb_gc_concurrency = 5") // sets auto concurrency and concurrency
-	result = tk.MustQuery("SELECT variable_value FROM mysql.tidb WHERE variable_name = 'tikv_gc_auto_concurrency'")
-	result.Check(testkit.Rows("false"))
-	result = tk.MustQuery("SELECT variable_value FROM mysql.tidb WHERE variable_name = 'tikv_gc_concurrency'")
-	result.Check(testkit.Rows("5"))
-
-	tk.MustExec("UPDATE mysql.tidb SET variable_value = 'true' WHERE variable_name='tikv_gc_auto_concurrency'")
-	result = tk.MustQuery("SELECT @@tidb_gc_concurrency;")
-	result.Check(testkit.Rows("-1")) // because auto_concurrency is turned on it takes precedence
-
-	tk.MustExec("REPLACE INTO mysql.tidb (variable_value, variable_name) VALUES ('15m', 'tikv_gc_run_interval')")
-	result = tk.MustQuery("SELECT @@GLOBAL.tidb_gc_run_interval;")
-	result.Check(testkit.Rows("15m0s"))
-	result = tk.MustQuery("SHOW GLOBAL VARIABLES LIKE 'tidb_gc_run_interval'")
-	result.Check(testkit.Rows("tidb_gc_run_interval 15m0s"))
-
-	tk.MustExec("SET GLOBAL tidb_gc_run_interval = '9m'") // too small
-	tk.MustQuery("SHOW WARNINGS").Check(testkit.Rows("Warning 1292 Truncated incorrect tidb_gc_run_interval value: '9m'"))
-	result = tk.MustQuery("SHOW GLOBAL VARIABLES LIKE 'tidb_gc_run_interval'")
-	result.Check(testkit.Rows("tidb_gc_run_interval 10m0s"))
-
-	tk.MustExec("SET GLOBAL tidb_gc_run_interval = '700000000000ns'") // specified in ns, also valid
-
-	_, err := tk.Exec("SET GLOBAL tidb_gc_run_interval = '11mins'")
-	require.Equal(t, "[variable:1232]Incorrect argument type to variable 'tidb_gc_run_interval'", err.Error())
 }
 
 func TestGlobalVarCollationServer(t *testing.T) {
@@ -1876,31 +1190,6 @@ func TestHostLengthMax(t *testing.T) {
 	tk.MustGetErrMsg(fmt.Sprintf(`CREATE USER 'abcddfjakldfjaldddds'@'%s'`, host2), "[ddl:1470]String 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa' is too long for host name (should be no longer than 255)")
 }
 
-func TestCommitRetryCount(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("use test")
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec("use test")
-
-	tk1.MustExec("create table no_retry (id int)")
-	tk1.MustExec("insert into no_retry values (1)")
-	tk1.MustExec("set @@tidb_retry_limit = 0")
-
-	tk1.MustExec("begin")
-	tk1.MustExec("update no_retry set id = 2")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update no_retry set id = 3")
-	tk2.MustExec("commit")
-
-	// No auto retry because retry limit is set to 0.
-	require.Error(t, tk1.ExecToErr("commit"))
-}
-
 func TestEnablePartition(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
@@ -1975,66 +1264,6 @@ func TestRollbackOnCompileError(t *testing.T) {
 		}
 	}
 	require.True(t, recoverErr)
-}
-
-func TestCastTimeToDate(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set time_zone = '-8:00'")
-	date := time.Now().In(time.FixedZone("", -8*int(time.Hour/time.Second)))
-	tk.MustQuery("select cast(time('12:23:34') as date)").Check(testkit.Rows(date.Format(time.DateOnly)))
-
-	tk.MustExec("set time_zone = '+08:00'")
-	date = time.Now().In(time.FixedZone("", 8*int(time.Hour/time.Second)))
-	tk.MustQuery("select cast(time('12:23:34') as date)").Check(testkit.Rows(date.Format(time.DateOnly)))
-}
-
-func TestSetGlobalTZ(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("set time_zone = '+08:00'")
-	tk.MustQuery("show variables like 'time_zone'").Check(testkit.Rows("time_zone +08:00"))
-
-	tk.MustExec("set global time_zone = '+00:00'")
-
-	tk.MustQuery("show variables like 'time_zone'").Check(testkit.Rows("time_zone +08:00"))
-
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustQuery("show variables like 'time_zone'").Check(testkit.Rows("time_zone +00:00"))
-}
-
-func TestErrorRollback(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t_rollback")
-	tk.MustExec("create table t_rollback (c1 int, c2 int, primary key(c1))")
-	tk.MustExec("insert into t_rollback values (0, 0)")
-
-	var wg sync.WaitGroup
-	cnt := 4
-	wg.Add(cnt)
-	num := 20
-
-	for i := 0; i < cnt; i++ {
-		go func() {
-			defer wg.Done()
-			tk := testkit.NewTestKit(t, store)
-			tk.MustExec("use test")
-			tk.MustExec("set @@session.tidb_retry_limit = 100")
-			for j := 0; j < num; j++ {
-				_, _ = tk.Exec("insert into t_rollback values (1, 1)")
-				tk.MustExec("update t_rollback set c2 = c2 + 1 where c1 = 0")
-			}
-		}()
-	}
-
-	wg.Wait()
-	tk.MustQuery("select c2 from t_rollback where c1 = 0").Check(testkit.Rows(fmt.Sprint(cnt * num)))
 }
 
 func TestDeletePanic(t *testing.T) {
@@ -2979,328 +2208,6 @@ func TestRowLock(t *testing.T) {
 	tk1.MustExec("commit")
 }
 
-// TestAutocommit . See https://dev.mysql.com/doc/internals/en/status-flags.html
-func TestAutocommit(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	tk.MustExec("drop table if exists t;")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-	tk.MustExec("create table t (id BIGINT PRIMARY KEY AUTO_INCREMENT NOT NULL)")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-	tk.MustExec("insert t values ()")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-	tk.MustExec("begin")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-	tk.MustExec("insert t values ()")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-	tk.MustExec("drop table if exists t")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-
-	tk.MustExec("create table t (id BIGINT PRIMARY KEY AUTO_INCREMENT NOT NULL)")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-	tk.MustExec("set autocommit=0")
-	require.Equal(t, 0, int(tk.Session().Status()&mysql.ServerStatusAutocommit))
-	tk.MustExec("insert t values ()")
-	require.Equal(t, 0, int(tk.Session().Status()&mysql.ServerStatusAutocommit))
-	tk.MustExec("commit")
-	require.Equal(t, 0, int(tk.Session().Status()&mysql.ServerStatusAutocommit))
-	tk.MustExec("drop table if exists t")
-	require.Equal(t, 0, int(tk.Session().Status()&mysql.ServerStatusAutocommit))
-	tk.MustExec("set autocommit='On'")
-	require.Greater(t, int(tk.Session().Status()&mysql.ServerStatusAutocommit), 0)
-
-	// When autocommit is 0, transaction start ts should be the first *valid*
-	// statement, rather than *any* statement.
-	tk.MustExec("create table t (id int key)")
-	tk.MustExec("set @@autocommit = 0")
-	tk.MustExec("rollback")
-	tk.MustExec("set @@autocommit = 0")
-
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("use test")
-	tk1.MustExec("insert into t select 1")
-	//nolint:all_revive,revive
-	tk.MustQuery("select * from t").Check(testkit.Rows("1"))
-	tk.MustExec("delete from t")
-
-	// When the transaction is rolled back, the global set statement would succeed.
-	tk.MustExec("set @@global.autocommit = 0")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (1)")
-	tk.MustExec("set @@global.autocommit = 1")
-	tk.MustExec("rollback")
-	tk.MustQuery("select count(*) from t where id = 1").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@global.autocommit").Check(testkit.Rows("1"))
-
-	// When the transaction is committed because of switching mode, the session set statement shold succeed.
-	tk.MustExec("set autocommit = 0")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (1)")
-	tk.MustExec("set autocommit = 1")
-	tk.MustExec("rollback")
-	tk.MustQuery("select count(*) from t where id = 1").Check(testkit.Rows("1"))
-	tk.MustQuery("select @@autocommit").Check(testkit.Rows("1"))
-
-	tk.MustExec("set autocommit = 0")
-	tk.MustExec("insert into t values (2)")
-	tk.MustExec("set autocommit = 1")
-	tk.MustExec("rollback")
-	tk.MustQuery("select count(*) from t where id = 2").Check(testkit.Rows("1"))
-	tk.MustQuery("select @@autocommit").Check(testkit.Rows("1"))
-
-	// Set should not take effect if the mode is not changed.
-	tk.MustExec("set autocommit = 0")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (3)")
-	tk.MustExec("set autocommit = 0")
-	tk.MustExec("rollback")
-	tk.MustQuery("select count(*) from t where id = 3").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@autocommit").Check(testkit.Rows("0"))
-
-	tk.MustExec("set autocommit = 1")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (4)")
-	tk.MustExec("set autocommit = 1")
-	tk.MustExec("rollback")
-	tk.MustQuery("select count(*) from t where id = 4").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@autocommit").Check(testkit.Rows("1"))
-}
-
-// TestTxnLazyInitialize tests that when autocommit = 0, not all statement starts
-// a new transaction.
-func TestTxnLazyInitialize(t *testing.T) {
-	testTxnLazyInitialize(t, false)
-	testTxnLazyInitialize(t, true)
-}
-
-func testTxnLazyInitialize(t *testing.T, isPessimistic bool) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (id int)")
-	if isPessimistic {
-		tk.MustExec("set tidb_txn_mode = 'pessimistic'")
-	}
-
-	tk.MustExec("set @@autocommit = 0")
-	_, err := tk.Session().Txn(true)
-	require.True(t, kv.ErrInvalidTxn.Equal(err))
-	txn, err := tk.Session().Txn(false)
-	require.NoError(t, err)
-	require.False(t, txn.Valid())
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-
-	// Those statements should not start a new transaction automatically.
-	tk.MustQuery("select 1")
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-
-	tk.MustExec("set @@tidb_general_log = 0")
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-
-	tk.MustQuery("explain select * from t")
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-
-	// Begin statement should start a new transaction.
-	tk.MustExec("begin")
-	txn, err = tk.Session().Txn(false)
-	require.NoError(t, err)
-	require.True(t, txn.Valid())
-	tk.MustExec("rollback")
-
-	tk.MustExec("select * from t")
-	txn, err = tk.Session().Txn(false)
-	require.NoError(t, err)
-	require.True(t, txn.Valid())
-	tk.MustExec("rollback")
-
-	tk.MustExec("insert into t values (1)")
-	txn, err = tk.Session().Txn(false)
-	require.NoError(t, err)
-	require.True(t, txn.Valid())
-	tk.MustExec("rollback")
-}
-
-func TestGlobalVarAccessor(t *testing.T) {
-	varName := "max_allowed_packet"
-	varValue := strconv.FormatUint(variable.DefMaxAllowedPacket, 10) // This is the default value for max_allowed_packet
-
-	// The value of max_allowed_packet should be a multiple of 1024,
-	// so the setting of varValue1 and varValue2 would be truncated to varValue0
-	varValue0 := "4194304"
-	varValue1 := "4194305"
-	varValue2 := "4194306"
-
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-
-	se := tk.Session().(variable.GlobalVarAccessor)
-	// Get globalSysVar twice and get the same value
-	v, err := se.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, varValue, v)
-	v, err = se.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, varValue, v)
-	// Set global var to another value
-	err = se.SetGlobalSysVar(context.Background(), varName, varValue1)
-	require.NoError(t, err)
-	v, err = se.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, varValue0, v)
-	require.NoError(t, tk.Session().CommitTxn(context.TODO()))
-
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("use test")
-	se1 := tk1.Session().(variable.GlobalVarAccessor)
-	v, err = se1.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, varValue0, v)
-	err = se1.SetGlobalSysVar(context.Background(), varName, varValue2)
-	require.NoError(t, err)
-	v, err = se1.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, varValue0, v)
-	require.NoError(t, tk1.Session().CommitTxn(context.TODO()))
-
-	// Make sure the change is visible to any client that accesses that global variable.
-	v, err = se.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, varValue0, v)
-
-	// For issue 10955, make sure the new session load `max_execution_time` into sessionVars.
-	tk1.MustExec("set @@global.max_execution_time = 100")
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec("use test")
-	require.Equal(t, uint64(100), tk2.Session().GetSessionVars().MaxExecutionTime)
-	tk1.MustExec("set @@global.max_execution_time = 0")
-
-	result := tk.MustQuery("show global variables  where variable_name='sql_select_limit';")
-	result.Check(testkit.Rows("sql_select_limit 18446744073709551615"))
-	result = tk.MustQuery("show session variables  where variable_name='sql_select_limit';")
-	result.Check(testkit.Rows("sql_select_limit 18446744073709551615"))
-	tk.MustExec("set session sql_select_limit=100000000000;")
-	result = tk.MustQuery("show global variables where variable_name='sql_select_limit';")
-	result.Check(testkit.Rows("sql_select_limit 18446744073709551615"))
-	result = tk.MustQuery("show session variables where variable_name='sql_select_limit';")
-	result.Check(testkit.Rows("sql_select_limit 100000000000"))
-	tk.MustExec("set @@global.sql_select_limit = 1")
-	result = tk.MustQuery("show global variables where variable_name='sql_select_limit';")
-	result.Check(testkit.Rows("sql_select_limit 1"))
-	tk.MustExec("set @@global.sql_select_limit = default")
-	result = tk.MustQuery("show global variables where variable_name='sql_select_limit';")
-	result.Check(testkit.Rows("sql_select_limit 18446744073709551615"))
-
-	result = tk.MustQuery("select @@global.autocommit;")
-	result.Check(testkit.Rows("1"))
-	result = tk.MustQuery("select @@autocommit;")
-	result.Check(testkit.Rows("1"))
-	tk.MustExec("set @@global.autocommit = 0;")
-	result = tk.MustQuery("select @@global.autocommit;")
-	result.Check(testkit.Rows("0"))
-	result = tk.MustQuery("select @@autocommit;")
-	result.Check(testkit.Rows("1"))
-	tk.MustExec("set @@global.autocommit=1")
-
-	err = tk.ExecToErr("set global time_zone = 'timezone'")
-	require.Error(t, err)
-	require.True(t, terror.ErrorEqual(err, variable.ErrUnknownTimeZone))
-}
-
-func TestUpgradeSysvars(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	se := tk.Session().(variable.GlobalVarAccessor)
-
-	// Set the global var to a non-canonical form of the value
-	// i.e. implying that it was set from an earlier version of TiDB.
-
-	tk.MustExec(`REPLACE INTO mysql.global_variables (variable_name, variable_value) VALUES ('tidb_enable_noop_functions', '0')`)
-	domain.GetDomain(tk.Session()).NotifyUpdateSysVarCache(true) // update cache
-	v, err := se.GetGlobalSysVar("tidb_enable_noop_functions")
-	require.NoError(t, err)
-	require.Equal(t, "OFF", v)
-
-	// Set the global var to ""  which is the invalid version of this from TiDB 4.0.16
-	// the err is quashed by the GetGlobalSysVar, and the default value is restored.
-	// This helps callers of GetGlobalSysVar(), which can't individually be expected
-	// to handle upgrade/downgrade issues correctly.
-
-	tk.MustExec(`REPLACE INTO mysql.global_variables (variable_name, variable_value) VALUES ('rpl_semi_sync_slave_enabled', '')`)
-	domain.GetDomain(tk.Session()).NotifyUpdateSysVarCache(true) // update cache
-	v, err = se.GetGlobalSysVar("rpl_semi_sync_slave_enabled")
-	require.NoError(t, err)
-	require.Equal(t, "OFF", v) // the default value is restored.
-	result := tk.MustQuery("SHOW VARIABLES LIKE 'rpl_semi_sync_slave_enabled'")
-	result.Check(testkit.Rows("rpl_semi_sync_slave_enabled OFF"))
-
-	// Ensure variable out of range is converted to in range after upgrade.
-	// This further helps for https://github.com/pingcap/tidb/pull/28842
-
-	tk.MustExec(`REPLACE INTO mysql.global_variables (variable_name, variable_value) VALUES ('tidb_executor_concurrency', '999')`)
-	domain.GetDomain(tk.Session()).NotifyUpdateSysVarCache(true) // update cache
-	v, err = se.GetGlobalSysVar("tidb_executor_concurrency")
-	require.NoError(t, err)
-	require.Equal(t, "256", v) // the max value is restored.
-
-	// Handle the case of a completely bogus value from an earlier version of TiDB.
-	// This could be the case if an ENUM sysvar removes a value.
-
-	tk.MustExec(`REPLACE INTO mysql.global_variables (variable_name, variable_value) VALUES ('tidb_enable_noop_functions', 'SOMEVAL')`)
-	domain.GetDomain(tk.Session()).NotifyUpdateSysVarCache(true) // update cache
-	v, err = se.GetGlobalSysVar("tidb_enable_noop_functions")
-	require.NoError(t, err)
-	require.Equal(t, "OFF", v) // the default value is restored.
-}
-
-func TestSetInstanceSysvarBySetGlobalSysVar(t *testing.T) {
-	varName := "tidb_general_log"
-	defaultValue := "OFF" // This is the default value for tidb_general_log
-
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	se := tk.Session().(variable.GlobalVarAccessor)
-
-	// Get globalSysVar twice and get the same default value
-	v, err := se.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, defaultValue, v)
-	v, err = se.GetGlobalSysVar(varName)
-	require.NoError(t, err)
-	require.Equal(t, defaultValue, v)
-
-	// session.GetGlobalSysVar would not get the value which session.SetGlobalSysVar writes,
-	// because SetGlobalSysVar calls SetGlobalFromHook, which uses TiDBGeneralLog's SetGlobal,
-	// but GetGlobalSysVar could not access TiDBGeneralLog's GetGlobal.
-
-	// set to "1"
-	err = se.SetGlobalSysVar(context.Background(), varName, "ON")
-	require.NoError(t, err)
-	v, err = se.GetGlobalSysVar(varName)
-	tk.MustQuery("select @@global.tidb_general_log").Check(testkit.Rows("1"))
-	require.NoError(t, err)
-	require.Equal(t, defaultValue, v)
-
-	// set back to "0"
-	err = se.SetGlobalSysVar(context.Background(), varName, defaultValue)
-	require.NoError(t, err)
-	v, err = se.GetGlobalSysVar(varName)
-	tk.MustQuery("select @@global.tidb_general_log").Check(testkit.Rows("0"))
-	require.NoError(t, err)
-	require.Equal(t, defaultValue, v)
-}
-
 func TestMatchIdentity(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
@@ -3374,53 +2281,6 @@ func TestGetSysVariables(t *testing.T) {
 	tk.MustExec("select @@session.performance_schema_max_mutex_classes")
 	tk.MustExec("select @@local.performance_schema_max_mutex_classes")
 	tk.MustGetErrMsg("select @@global.last_insert_id", "[variable:1238]Variable 'last_insert_id' is a SESSION variable")
-}
-
-// TestInTrans . See https://dev.mysql.com/doc/internals/en/status-flags.html
-func TestInTrans(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t (id BIGINT PRIMARY KEY AUTO_INCREMENT NOT NULL)")
-	tk.MustExec("insert t values ()")
-	tk.MustExec("begin")
-	txn, err := tk.Session().Txn(true)
-	require.NoError(t, err)
-	require.True(t, txn.Valid())
-	tk.MustExec("insert t values ()")
-	require.True(t, txn.Valid())
-	tk.MustExec("drop table if exists t;")
-	require.False(t, txn.Valid())
-	tk.MustExec("create table t (id BIGINT PRIMARY KEY AUTO_INCREMENT NOT NULL)")
-	require.False(t, txn.Valid())
-	tk.MustExec("insert t values ()")
-	require.False(t, txn.Valid())
-	tk.MustExec("commit")
-	tk.MustExec("insert t values ()")
-
-	tk.MustExec("set autocommit=0")
-	tk.MustExec("begin")
-	require.True(t, txn.Valid())
-	tk.MustExec("insert t values ()")
-	require.True(t, txn.Valid())
-	tk.MustExec("commit")
-	require.False(t, txn.Valid())
-	tk.MustExec("insert t values ()")
-	require.True(t, txn.Valid())
-	tk.MustExec("commit")
-	require.False(t, txn.Valid())
-
-	tk.MustExec("set autocommit=1")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (id BIGINT PRIMARY KEY AUTO_INCREMENT NOT NULL)")
-	tk.MustExec("begin")
-	require.True(t, txn.Valid())
-	tk.MustExec("insert t values ()")
-	require.True(t, txn.Valid())
-	tk.MustExec("rollback")
-	require.False(t, txn.Valid())
 }
 
 func TestSession(t *testing.T) {
