+++ session/sessiontest/vars/BUILD.bazel
@@ -1,44 +1,29 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
-    name = "sessiontest_test",
+    name = "vars_test",
     timeout = "short",
     srcs = [
         "main_test.go",
-        "session_test.go",
+        "vars_test.go",
     ],
     flaky = True,
-    race = "on",
-    shard_count = 50,
+    shard_count = 11,
     deps = [
         "//config",
         "//domain",
-        "//expression",
+        "//errno",
         "//kv",
-        "//parser/ast",
-        "//parser/auth",
-        "//parser/format",
-        "//parser/model",
         "//parser/mysql",
         "//parser/terror",
-        "//privilege/privileges",
-        "//session",
-        "//sessionctx",
         "//sessionctx/variable",
-        "//store/copr",
-        "//store/mockstore",
-        "//table/tables",
         "//testkit",
         "//testkit/testmain",
         "//testkit/testsetup",
-        "//testkit/testutil",
-        "//types",
-        "//util",
-        "//util/memory",
-        "//util/sqlexec",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//tikv",
+        "@com_github_tikv_client_go_v2//txnkv/transaction",
         "@org_uber_go_goleak//:goleak",
     ],
 )
